+++ worker/main.go
@@ -8,8 +8,8 @@ import (
 	"github.com/go-redis/redis/v7"
 	"github.com/google/uuid"
 	"github.com/gorilla/websocket"
+	"go.uber.org/zap"
 	"google.golang.org/protobuf/proto"
-	"log"
 	"net/http"
 	"os"
 	"os/signal"
@@ -38,22 +38,44 @@ var connections = connectionsState{
 	state: make(map[string]*SockConnection),
 }
 
+var options *common.DSockOptions
+var logger *zap.Logger
 var redisClient *redis.Client
-var options common.DSockOptions
 
 func init() {
-	options = common.GetOptions()
+	var err error
+
+	options, err = common.GetOptions()
+
+	if err != nil {
+		println("Could not get options. Make sure your config is valid!")
+		panic(err)
+	}
+
+	if options.Debug {
+		logger, err = zap.NewDevelopment()
+	} else {
+		logger, err = zap.NewProduction()
+	}
+
+	if err != nil {
+		println("Could not create logger")
+		panic(err)
+	}
 }
 
 func main() {
-	log.Printf("Starting dSock worker %s\n", common.DSockVersion)
-
+	logger.Info("Starting dSock worker",
+		zap.String("version", common.DSockVersion),
+	)
 	// Setup application
 	redisClient = redis.NewClient(options.RedisOptions)
 
 	_, err := redisClient.Ping().Result()
 	if err != nil {
-		panic(err)
+		logger.Error("Could not connect to Redis (ping)",
+			zap.Error(err),
+		)
 	}
 
 	messageSubscription := redisClient.Subscribe(workerId)
@@ -65,7 +87,9 @@ func main() {
 		gin.SetMode(gin.ReleaseMode)
 	}
 
-	router := gin.Default()
+	router := common.NewGinEngine(logger, options)
+	router.Use(common.RequestIdMiddleware)
+
 	router.GET(common.PathConnect, connectHandler)
 
 	// Start HTTP server
@@ -78,7 +102,9 @@ func main() {
 
 	go func() {
 		if err := srv.ListenAndServe(); err != nil && err != http.ErrServerClosed {
-			log.Printf("Failed listening: %s\n", err)
+			logger.Error("Failed listening",
+				zap.Error(err),
+			)
 			options.QuitChannel <- struct{}{}
 		}
 	}()
@@ -89,7 +115,9 @@ func main() {
 			redisMessage, err := messageSubscription.ReceiveMessage()
 			if err != nil {
 				// TODO: Possibly add better handling
-				log.Printf("Error receiving from Redis: %s\n", err)
+				logger.Error("Error receiving message from Redis",
+					zap.Error(err),
+				)
 				break
 			}
 
@@ -99,7 +127,9 @@ func main() {
 
 			if err != nil {
 				// Couldn't parse message
-				log.Printf("Invalid message from Redis: %s\n", err)
+				logger.Error("Invalid message received from Redis",
+					zap.Error(err),
+				)
 				continue
 			}
 
@@ -117,7 +147,9 @@ func main() {
 			redisMessage, err := channelSubscription.ReceiveMessage()
 			if err != nil {
 				// TODO: Possibly add better handling
-				log.Printf("Error receiving from Redis: %s\n", err)
+				logger.Error("Error receiving message from Redis",
+					zap.Error(err),
+				)
 				break
 			}
 
@@ -127,7 +159,9 @@ func main() {
 
 			if err != nil {
 				// Couldn't parse channel action
-				log.Printf("Invalid channel action from Redis: %s\n", err)
+				logger.Error("Invalid message received from Redis",
+					zap.Error(err),
+				)
 				continue
 			}
 
@@ -139,7 +173,9 @@ func main() {
 		}
 	}()
 
-	log.Printf("Starting on: %s\n", options.Address)
+	logger.Info("Listening",
+		zap.String("address", options.Address),
+	)
 
 	// Listen for signal or message in quit channel
 	signal.Notify(signalQuit, syscall.SIGINT, syscall.SIGTERM)
@@ -152,12 +188,14 @@ func main() {
 	signalQuit = nil
 
 	// Server shutdown
-	log.Print("Shutting server down...\n")
+	logger.Info("Shutting down")
 
 	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
 	defer cancel()
 	if err := srv.Shutdown(ctx); err != nil {
-		log.Printf("Error during server shutdown: %v\n", err)
+		logger.Error("Error during server shutdown",
+			zap.Error(err),
+		)
 	}
 
 	// Cleanup
@@ -172,5 +210,6 @@ func main() {
 	// Allow time to disconnect & clear from Redis
 	time.Sleep(time.Second)
 
-	log.Println("Server stopped")
+	logger.Info("Stopped")
+	_ = logger.Sync()
 }